LayoutTests/accessibility/mac/select-text-stays-inside-text-controls-expected.txt


 outside text
This tests that the select text mechanism will not return results outside of a text input control.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS result is 'textfield'
PASS document.getElementById('textfield').selectionStart is 0
PASS document.getElementById('textfield').selectionEnd is 9
PASS result is ''
PASS document.getElementById('textfield').selectionStart is 0
PASS document.getElementById('textfield').selectionEnd is 0
PASS result is 'text area'
PASS document.getElementById('textarea').selectionStart is 0
PASS document.getElementById('textarea').selectionEnd is 9
PASS result is ''
PASS document.getElementById('textarea').selectionStart is 0
PASS document.getElementById('textarea').selectionEnd is 0
PASS successfullyParsed is true

TEST COMPLETE
